Refactored AuthService a bit
