+++ social-app-backend/src/main/java/razepl/dev/socialappbackend/config/interfaces/JwtServiceInterface.java
@@ -2,8 +2,6 @@ package razepl.dev.socialappbackend.config.interfaces;
 
 import io.jsonwebtoken.Claims;
 import org.springframework.security.core.userdetails.UserDetails;
-import razepl.dev.socialappbackend.jwt.interfaces.Token;
-import razepl.dev.socialappbackend.user.User;
 
 import java.util.Map;
 import java.util.function.Function;
@@ -18,6 +16,4 @@ public interface JwtServiceInterface {
     String generateToken(Map<String, Object> additionalClaims, UserDetails userDetails);
 
     boolean isTokenValid(String jwtToken, UserDetails userDetails);
-
-    Token buildToken(String jwtToken, User user);
 }
