+++ social-app-backend/src/main/java/razepl/dev/socialappbackend/jwt/interfaces/TokenManager.java
@@ -1,0 +1,10 @@
+package razepl.dev.socialappbackend.jwt.interfaces;
+
+import razepl.dev.socialappbackend.auth.apicalls.AuthResponse;
+import razepl.dev.socialappbackend.user.User;
+
+public interface TokenManager {
+    String buildUsersAuthToken(User user);
+
+    AuthResponse buildTokensIntoResponse(String authToken);
+}
