+++ social-app-backend/src/main/resources/application.properties
@@ -5,7 +5,6 @@ spring.jpa.hibernate.ddl-auto=create-drop
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.properties.hibernate.format_sql=true
-
 security.jwt.expiration-time=86400000
 security.jwt.refresh-time=604800000
 security.jwt.encoding-key=25442A472D4B6150645367566B597033733676397924423F4528482B4D625165
