+++ social-app-backend/src/main/resources/logback.xml
@@ -6,6 +6,6 @@
     </appender>
 
     <root level="debug">
-        <appender-ref ref="STDOUT" />
+        <appender-ref ref="STDOUT"/>
     </root>
 </configuration>
