+++ social-app-backend/src/test/java/razepl/dev/socialappbackend/exceptions/JsonMapperException.java
@@ -1,6 +1,6 @@
 package razepl.dev.socialappbackend.exceptions;
 
-public class JsonMapperException extends RuntimeException{
+public class JsonMapperException extends RuntimeException {
     public JsonMapperException(String message) {
         super(message);
     }
